+++ R/scale_hdx.R
@@ -51,11 +51,12 @@
 #'     along with other styling.
 #'
 #' @export
-scale_color_hdx_discrete <- function(...) {
+scale_color_hdx_discrete <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_discrete(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
@@ -66,11 +67,12 @@ scale_colour_hdx_discrete <- scale_color_hdx_discrete
 
 #' @rdname scale_hdx
 #' @export
-scale_color_hdx_gray <- function(...) {
+scale_color_hdx_gray <- function(na.value = hdx_hex("tomato-hdx"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_gray(),
-    na.value = hdx_hex("tomato-hdx"),
+    na.value = na.value,
     ...
   )
 }
@@ -89,11 +91,12 @@ scale_color_hdx_grey <- scale_color_hdx_gray
 
 #' @rdname scale_hdx
 #' @export
-scale_color_hdx_mint <- function(...) {
+scale_color_hdx_mint <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_mint(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
@@ -104,11 +107,12 @@ scale_colour_hdx_mint <- scale_color_hdx_mint
 
 #' @rdname scale_hdx
 #' @export
-scale_color_hdx_sapphire <- function(...) {
+scale_color_hdx_sapphire <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetic = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_sapphire(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
@@ -119,11 +123,12 @@ scale_colour_hdx_sapphire <- scale_color_hdx_sapphire
 
 #' @rdname scale_hdx
 #' @export
-scale_color_hdx_tomato <- function(...) {
+scale_color_hdx_tomato <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "colour",
+    scale_name = scale_name(),
     palette = hdx_pal_tomato(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
@@ -134,22 +139,24 @@ scale_colour_hdx_tomato <- scale_color_hdx_tomato
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_hdx_discrete <- function(...) {
+scale_fill_hdx_discrete <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_discrete(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_hdx_gray <- function(...) {
+scale_fill_hdx_gray <- function(na.value = hdx_hex("tomato-hdx"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_gray(),
-    na.value = hdx_hex("tomato-hdx"),
+    na.value = na.value,
     ...
   )
 }
@@ -160,103 +167,116 @@ scale_fill_hdx_grey <- scale_fill_hdx_gray
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_hdx_mint <- function(...) {
+scale_fill_hdx_mint <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_mint(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_hdx_sapphire <- function(...) {
+scale_fill_hdx_sapphire <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_sapphire(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_hdx_tomato <- function(...) {
+scale_fill_hdx_tomato <- function(na.value = hdx_hex("gray-light"), ...) {
   ggplot2::discrete_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = hdx_pal_tomato(),
-    na.value = hdx_hex("gray-light"),
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_gradient_hdx <- function(...) {
+scale_fill_gradient_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_gradient_hdx_sapphire <- function(...) {
+scale_fill_gradient_hdx_sapphire <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_gradient_hdx_mint <- function(...) {
+scale_fill_gradient_hdx_mint <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("mint-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_gradient_hdx_tomato <- function(...) {
+scale_fill_gradient_hdx_tomato <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("tomato-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
 
 #' @rdname scale_hdx
 #' @export
-scale_color_gradient_hdx <- function(...) {
+scale_color_gradient_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
@@ -267,14 +287,16 @@ scale_colour_gradient_hdx <- scale_color_gradient_hdx
 
 #' @rdname scale_hdx
 #' @export
-scale_color_gradient_hdx_sapphire <- function(...) {
+scale_color_gradient_hdx_sapphire <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
@@ -285,14 +307,16 @@ scale_colour_gradient_hdx_sapphire <- scale_color_gradient_hdx_sapphire
 
 #' @rdname scale_hdx
 #' @export
-scale_color_gradient_hdx_mint <- function(...) {
+scale_color_gradient_hdx_mint <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("mint-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
@@ -303,14 +327,16 @@ scale_colour_gradient_hdx_mint <- scale_color_gradient_hdx_mint
 
 #' @rdname scale_hdx
 #' @export
-scale_color_gradient_hdx_tomato <- function(...) {
+scale_color_gradient_hdx_tomato <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::seq_gradient_pal(
       low = "white",
       high = hdx_hex("tomato-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
@@ -321,15 +347,17 @@ scale_colour_gradient_hdx_tomato <- scale_color_gradient_hdx_tomato
 
 #' @rdname scale_hdx
 #' @export
-scale_color_gradient2_hdx <- function(...) {
+scale_color_gradient2_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "color",
+    scale_name = scale_name(),
     palette = scales::div_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       mid = hdx_hex("gray-white"),
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }
@@ -340,15 +368,17 @@ scale_colour_gradient2_hdx <- scale_color_gradient2_hdx
 
 #' @rdname scale_hdx
 #' @export
-scale_fill_gradient2_hdx <- function(...) {
+scale_fill_gradient2_hdx <- function(na.value = "transparent", ...) {
   ggplot2::continuous_scale(
     aesthetics = "fill",
+    scale_name = scale_name(),
     palette = scales::div_gradient_pal(
       low = hdx_hex("tomato-hdx"),
       mid = hdx_hex("gray-white"),
       high = hdx_hex("sapphire-hdx")
     ),
     guide = "colorbar",
+    na.value = na.value,
     ...
   )
 }