+++ NEWS.md
@@ -1,3 +1,8 @@
+# Development
+
+- Implement `label_key_figures()` function to improve numeric scales labeling
+to match the data visualization guidelines.
+
 # gghdx 0.1.2
 
 * Use anonymous function format `function()` in `hdx_display_pal()`.