+++ NEWS.md
@@ -1,3 +1,5 @@
+# gghdx 0.1.3
+
 # Development
 
 - Implement `label_number_hdx()` function to improve numeric scales labeling